% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MQRE.R
\name{MQRE_function}
\alias{MQRE_function}
\title{The function provides the estimates based on M-quantile random effects regression.}
\usage{
MQRE_function(formula, saind, qtl, data, tol = 1e-04, maxit = 100,
  k = 1.345, k_v = 1.345)
}
\arguments{
\item{formula}{An object of class "formula"}

\item{saind}{Small area indicator (specification of the grouping)}

\item{qtl}{Define the M-quantile}

\item{data}{Data frame containing the variables in the model}

\item{tol}{Tolerance}

\item{maxit}{Maximum number of iterations}

\item{k}{Tuning constant}

\item{k_v}{Tuning constant for the estimation of the random effects}
}
\value{
Returns a list including:
\itemize{
\item coefficients - Regression parameters
\item stde.beta - Standard error of the regression parameters
\item sigma2u - Varianz of the random effects (Level 2)
\item std.u - Standard error of the varianz of the random effects (Level 2)
\item sigma2e - Varianz of the error term (Level 1)
\item std.e - Standard error of the varianz of the error term (Level 1)
\item quantile - Estimated M-quantile
\item samplesize - Sample size in each group
\item iterations - Number of iterations in the fixed point algorithm
\item Kvalue - Value of the K matrix in the estimation equations
\item rand.eff - Vector of the random effects
}
}
\description{
The function provides the estimates based on M-quantile random effects regression.
}
\examples{
# Data generation
library(saeSim)
set.seed(1)
groups<-30 # Define the number of groups
sample_size<-sample(x=5:15,size=groups,replace=TRUE) # Define the sample sizes
setup<-sim_base(data = base_id(nDomains=groups,nUnits=sample_size)) \%>\%
 sim_gen_x(mean=0,sd=10)       \%>\%
 sim_gen_v(mean=0,sd=300)         \%>\%
 sim_gen_e(mean=0,sd=800)         \%>\%
 sim_resp_eq(y = 4500 - 400*x + v + e)
 # Generate 10 samples
Pop<-sim(setup,R=10)
sample_data<-Pop[[1]] # Select the first sample for the test

# Fit the MQRE model
fit<-MQRE_function(formula=y~x,saind=sample_data$idD,data=sample_data,qtl=0.4)
}
\references{
Tzavidis, N.; Salvati, N.; Schmid, T.; Flouri, E. and Midouhas, E.: Longitudinal analysis of the Strengths and Difficulties Questionnaire scores of the Millennium Cohort Study children in England using M-quantile random effects regression, Journal of the Royal Statistical Society: Series A, accepted.
}

